fix(checkoutv2): display product item form checkout/cart page
