+++ apps/store/src/app/[locale]/new/checkout/CheckoutForm.tsx
@@ -1,10 +1,10 @@
 'use client'
 
 import { useApolloClient } from '@apollo/client'
-import { useRouter } from 'next/navigation'
 import { useTranslation } from 'next-i18next'
-import { useCallback, useEffect, useRef, useState, useTransition } from 'react'
+import { useState } from 'react'
 import { Text, Button, BankIdIcon, yStack } from 'ui'
+import { useAsyncRouterPush } from '@/appComponents/useAsyncRouterPush'
 import { useHandleSubmitCheckout } from '@/components/CheckoutPage/useHandleSubmitCheckout'
 import * as FullscreenDialog from '@/components/FullscreenDialog/FullscreenDialog'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
@@ -69,9 +69,6 @@ export function CheckoutForm({
       resetShopSession()
 
       const nextCheckoutStep = getNextCheckoutStep(data.currentMember.paymentInformation.status)
-      // TODO: router from next/navigation has a different API so we can't wait for the router to be updated
-      // before we navigate to the next step. We probably gonna need to change how BankIdDialog works so
-      // we can have smooth transitions between sucess state and connect payment page.
       await push(getCheckoutStepLink({ locale, step: nextCheckoutStep, shopSessionId }))
     },
     onError() {
@@ -168,37 +165,4 @@ function getNextCheckoutStep(connectPaymentStatus: MemberPaymentConnectionStatus
   const nextCheckoutStep = checkoutSteps[1]
 
   return nextCheckoutStep
-}
-
-// Hacky solution to wait for the router to finish the transition as `router` from `next/navigation`
-// package doesn't have a promise based API to wait for the transition to finish. They don't expose
-// events either. They claim that all navigations in app router are built on React Transitions so
-// the way forward for knowing wherebouts of a navigation is to use `useTransition` hook from React.
-// More info https://github.com/vercel/next.js/discussions/41934#discussioncomment-8996669
-function useAsyncRouterPush() {
-  const router = useRouter()
-  // We assume no other transition will happening at the same time
-  const [isPending] = useTransition()
-  const promiseResolveFnRef = useRef<((value?: unknown) => void) | null>(null)
-
-  useEffect(() => {
-    if (!isPending) {
-      promiseResolveFnRef.current?.()
-    }
-  }, [isPending])
-
-  const push = useCallback(
-    (url: string) => {
-      return new Promise((resolve) => {
-        // Ideally we could attach the resolve function to a router event but they're not exposed
-        // by next/navigation router. So we're using a ref to store the resolve function and calling
-        // it when the transition is finished based on the `isPending` value.
-        promiseResolveFnRef.current = resolve
-        router.push(url)
-      })
-    },
-    [router],
-  )
-
-  return push
 }
