+++ apps/store/src/components/CartPage/CartPage.tsx
@@ -84,7 +84,7 @@ export const CartPage = () => {
                       offer={item}
                       defaultExpanded={router.query[QueryParam.ExpandCart] === '1'}
                     >
-                      <EditActionButton shopSessionId={shopSession.id} offer={item} />
+                      <EditActionButton offer={item} />
                       <RemoveActionButton
                         shopSessionId={shopSession.id}
                         offer={item}
