+++ apps/store/src/components/ProductItem/useEditProductOffer.ts
@@ -1,79 +1,51 @@
-import { useApolloClient } from '@apollo/client'
 import { datadogLogs } from '@datadog/browser-logs'
-import { useRouter } from 'next/router'
 import { useState } from 'react'
-import { useProductMetadata } from '@/components/LayoutWithMenu/productMetadataHooks'
+import { useAsyncRouterPush } from '@/appComponents/useAsyncRouterPush'
 import { OPEN_PRICE_CALCULATOR_QUERY_PARAM } from '@/components/ProductPage/PurchaseForm/useOpenPriceCalculatorQueryParam'
 import { PRELOADED_PRICE_INTENT_QUERY_PARAM } from '@/components/ProductPage/PurchaseForm/usePreloadedPriceIntentId'
 import { CART_ENTRY_TO_REPLACE_QUERY_PARAM } from '@/components/ProductPage/useCartEntryToReplace'
 import { useShowAppError } from '@/services/appErrors/appErrorAtom'
-import { getPriceTemplate } from '@/services/PriceCalculator/PriceCalculator.helpers'
-import { priceIntentServiceInitClientSide } from '@/services/priceIntent/PriceIntentService'
+import type { ProductOfferFragment } from '@/services/graphql/generated'
+import { Features } from '@/utils/Features'
 
 type State = 'idle' | 'loading' | 'error'
 
 type Params = {
-  shopSessionId: string
-  offerId: string
-  data: any
-  productName: string
+  offer: ProductOfferFragment
 }
 
 export const useEditProductOffer = () => {
   const [state, setState] = useState<State>('idle')
-  const apolloClient = useApolloClient()
-  const products = useProductMetadata()
-  const router = useRouter()
   const showError = useShowAppError()
+  const push = useAsyncRouterPush()
 
-  const editProductOffer = async (params: Params) => {
-    setState('loading')
-    // TODO: Check if product offer is connected to price intent -- if not, create new one
-
-    const product = products?.find((item) => item.name === params.productName)
-    if (!product) {
-      datadogLogs.logger.error('Edit Offer | Product not found', {
-        productName: params.productName,
-        productNames: products?.map((item) => item.name),
-        offerId: params.offerId,
-      })
-      throw new Error(`Edit Offer | Product not found: ${params.productName}`)
-    }
-
-    const priceIntentService = priceIntentServiceInitClientSide(apolloClient)
-    const priceTemplate = getPriceTemplate(params.productName)
-    if (!priceTemplate) {
-      setState('error')
-      datadogLogs.logger.error('Price template not found', { productName: params.productName })
-      return
-    }
-
+  const editProductOffer = async ({ offer }: Params) => {
     try {
-      const priceIntent = await priceIntentService.create({
-        shopSessionId: params.shopSessionId,
-        productName: params.productName,
-        priceTemplate,
-      })
-      await priceIntentService.update({
-        priceIntentId: priceIntent.id,
-        data: params.data,
-        customer: { shopSessionId: params.shopSessionId },
-      })
+      setState('loading')
 
-      await router.push({
-        pathname: product.pageLink,
-        query: {
-          [OPEN_PRICE_CALCULATOR_QUERY_PARAM]: 1,
-          [CART_ENTRY_TO_REPLACE_QUERY_PARAM]: params.offerId,
-          [PRELOADED_PRICE_INTENT_QUERY_PARAM]: priceIntent.id,
-        },
-      })
+      let targetUrl = new URL(offer.product.pageLink, window.location.origin)
+      if (
+        Features.enabled('PRICE_CALCULATOR_PAGE') &&
+        offer.product.priceCalculatorPageLink != null
+      ) {
+        targetUrl = new URL(offer.product.priceCalculatorPageLink, window.location.origin)
+      } else {
+        targetUrl.searchParams.set(OPEN_PRICE_CALCULATOR_QUERY_PARAM, '1')
+        // NOTE: Temporary code path, no need to handle this after PriceCalculatorPage is used everywhere
+        if (offer.priceIntentId == null) {
+          throw new Error('Expected to have offer.priceIntentId')
+        }
+        targetUrl.searchParams.set(PRELOADED_PRICE_INTENT_QUERY_PARAM, offer.priceIntentId)
+      }
+      targetUrl.searchParams.set(CART_ENTRY_TO_REPLACE_QUERY_PARAM, offer.id)
+      await push(targetUrl.toString())
+
       setState('idle')
     } catch (error) {
       setState('error')
       datadogLogs.logger.error('Error while editing product offer', {
         error,
-        offerId: params.offerId,
+        offerId: offer.id,
       })
       showError(error as Error)
     }
